I removed main_obprm and main_query from the testRigid and testSerial directories.  main_obprm and main_query are not in the /src dir, and you can control which CfgType to use through the Makefile.
BoundingBox problem:  There was a bounding box problem that I fixed, now the 
bounding box is caluclated properly, this problem came up in qry and 
reported and found by Luke.  This still is a major problem, in that qry will 
NEVER read the input form a mapfile, so any specific bounding box you 
specify during map creation will not be presereved when callilng qry.





git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/trunk@1200 5e2c2758-810a-0410-8eb3-8e4e92d4004e